Replace Shell.WindowTracker with Shell.AppSystem
The app_id cannot be retrieved for new applications with Shell.WindowTracker
on Wayland. It only works with applications that are already started
when the extension is started. Therefore, the defined applications
cannot enable Caffeine on Wayland.

Shell.AppSystem works on Wayland and X11, but it works differently. We
need to connect to each defined application and wait for a change in
it's state to decide if Caffeine should be enabled or not.